+++ chess/constants.py
