+++ chess/exceptions.py
