+++ chess/ui/animations.py
