+++ scripts/run_main.py
@@ -1,0 +1,4 @@
+from chess.ui.main import Main
+
+if __name__ == "__main__":
+    Main().mainloop()
