+++ scripts/run_server.py
@@ -1,0 +1,4 @@
+from chess.network import server
+
+if __name__ == "__main__":
+    server.server_program()
